Establish separate thumbnail collection.
